group33/workspace/requirement.txt
aiohttp==2.3.2
aiohttp-cors==0.5.3
async-timeout==2.0.0
bleach==1.5.0
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.9
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8
mpmath==1.0.0
multidict==3.3.2
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.2.0
numpy==1.13.3
opencv-python==3.3.0.10
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sympy==1.1.1
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc2
terminado==0.6
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.6
yarl==0.13.0
